refactor(stream): decompose StreamView runtime
